upgrade: setup custom reponse with all CRUD operations related to customer

+++ src/main/java/lk/autocare/vehicle_service_system/usecase/customer/CustomerUseCaseImpl.java
@@ -3,6 +3,7 @@ package lk.autocare.vehicle_service_system.usecase.customer;
 import lk.autocare.vehicle_service_system.domain.models.Customer;
 import lk.autocare.vehicle_service_system.domain.repositories.CustomerRepository;
 import lk.autocare.vehicle_service_system.GlobalExceptionHandler.ResourceNotFoundException;
+import lk.autocare.vehicle_service_system.web.customer.DTOs.CustomerResponseDTO;
 import lombok.RequiredArgsConstructor;
 
 import java.util.List;
@@ -21,29 +22,32 @@ public class CustomerUseCaseImpl implements  CustomerUseCase {
 
     //save new customer
     @Override
-    public void saveCustomer(Customer customer){
-        customerRepository.saveCustomer(customer);
+    public Customer saveCustomer(Customer customer){
+        return customerRepository.saveCustomer(customer);
     }
 
     //update customer
     @Override
-    public void updateCustomer(Long customerId, Customer customer){
+    public Customer updateCustomer(Long customerId, Customer customer){
 
         //check customer availability then throw error to exception handler
         customerRepository.findById(customerId)
                 .orElseThrow(() -> new ResourceNotFoundException("Customer not found" + " " + customerId));
-        //customer found then save in db
-        customerRepository.updateCustomer(customerId,customer);
+
+        //set values to domain repo to save in db
+        return customerRepository.updateCustomer(customerId,customer);
     }
 
     //delete customer
     @Override
-    public void deleteCustomer(Long customerId){
+    public Customer deleteCustomer(Long customerId){
 
         //check customer availability then throw error to exception handler
-        customerRepository.findById(customerId)
-                .orElseThrow(() -> new ResourceNotFoundException("Customer not found" + " " + customerId));
+        if(customerRepository.findById(customerId).isEmpty()){
+            throw new ResourceNotFoundException("Customer not found(usecase)" + " " + customerId);
+        }
 
-        customerRepository.deleteCustomer(customerId);
+        //set customer to domain repo to delete
+       return customerRepository.deleteCustomer(customerId);
     }
 }
